fix: remove verifier in favour of change hashes
