+++ packages/publish/src/bump.ts
@@ -1,21 +1,9 @@
-import { confirm, isCancel } from '@clack/prompts';
 import { consola } from 'consola';
-import fs from 'fs-extra';
-import stringify from 'json-stringify-pretty-compact';
-import latestVersion from 'latest-version';
-import PQueue from 'p-queue';
-import path from 'pathe';
 import colors from 'picocolors';
 import semver from 'semver';
 
 import { getChanged } from './changed';
-import type {
-	BumpFlags,
-	BumpObject,
-	ChangedList,
-	Context,
-	PackageJson,
-} from './types';
+import type { BumpFlags, BumpObject, ChangedList } from './types';
 import { mergeFlags } from './utils';
 
 export const isValidBumpArg = (bumpArg: string): boolean => {
@@ -66,56 +54,7 @@ export const bumpValue = (
 	return false;
 };
 
-export const verifyVersion = async (
-	pkg: BumpObject,
-	bumpArg: string
-): Promise<BumpObject> => {
-	let npmVersion: string | boolean;
-	const newPkg = pkg;
-
-	try {
-		// Get latest version from NPM registry and compare if bumped version is greater than NPM
-		npmVersion = await latestVersion(pkg.name);
-		if (semver.gt(pkg.bumpVersion as string, npmVersion)) {
-			return pkg;
-		}
-	} catch (error: any) {
-		// If package isn't published on NPM yet, revert bump
-		if (error.name === 'PackageNotFoundError') {
-			newPkg.bumpVersion = newPkg.version;
-		}
-
-		return newPkg;
-	}
-
-	// If failed, do not publish
-	newPkg.noPublish = true;
-	if (bumpArg === 'from-package') {
-		return newPkg;
-	}
-
-	throw new Error(
-		colors.red(
-			`${newPkg.name} version mismatch. Not publishing.\n- NPM: ${npmVersion}\n- Bumped version: ${pkg.bumpVersion}`
-		)
-	);
-};
-
-export const writeUpdate = async (pkg: BumpObject): Promise<void> => {
-	const pkgPath = path.join(pkg.path, 'package.json');
-	const pkgJson: PackageJson = await fs.readJson(pkgPath);
-	pkgJson.version = pkg.bumpVersion;
-	pkgJson.publishHash = pkg.hash;
-	await fs.writeFile(pkgPath, stringify(pkgJson));
-};
-
-const queue = new PQueue({ concurrency: 12 });
-
-export const bumpPackages = async (
-	diff: ChangedList,
-	config: Context,
-	version: string
-) => {
+export const bumpPackages = async (diff: ChangedList, version: string) => {
 	// Create bump objects with bumped version
 	const bumpObjects: BumpObject[] = [];
 	for (const pkg of diff) {
@@ -132,22 +71,9 @@ export const bumpPackages = async (
 		bumpObjects.push(bumpObject);
 	}
 
-	// Check if package versions are free of conflicts on NPM
-	let pendingObjects;
-	if (!config.noVerify) {
-		pendingObjects = [];
-		for (const pkg of bumpObjects) {
-			const newPkg = queue.add(() => verifyVersion(pkg, version));
-			pendingObjects.push(newPkg);
-		}
-	}
-	// Wait for queue to finish
-	const verifiedObjects = await Promise.all(pendingObjects ?? bumpObjects);
-	consola.info(colors.bold(colors.blue('All packages verified.')));
-
 	// Print out all new versions and prompt user to continue
 	let count = 0;
-	for (const pkg of verifiedObjects) {
+	for (const pkg of bumpObjects) {
 		if (!pkg) {
 			throw new Error('Package object is undefined.');
 		}
@@ -156,33 +82,16 @@ export const bumpPackages = async (
 		);
 		count += 1;
 	}
+	consola.info(colors.magenta(`${count} packages will be updated.`));
 
-	if (!config.yes) {
-		const yes = await confirm({ message: `Bump ${count} packages?` });
-		if (!yes || isCancel(yes)) {
-			throw new Error('Bump cancelled.');
-		}
-	} else {
-		consola.info(colors.bold(colors.blue(`Bumping ${count} packages...`)));
-	}
-
-	// Update all padkage.json files
-	for (const pkg of verifiedObjects) {
+	// Sanity check
+	for (const pkg of bumpObjects) {
 		if (!pkg) {
 			throw new Error('Package object is undefined.');
 		}
-
-		// Skip if noPublish flag is set
-		if (!pkg.noPublish) {
-			await writeUpdate(pkg);
-		} else {
-			consola.info(
-				colors.yellow(`Skipping ${pkg.name} as it is set to not publish.`)
-			);
-		}
 	}
 
-	return verifiedObjects as unknown as BumpObject[];
+	return bumpObjects;
 };
 
 export const bump = async (version: string, options: BumpFlags) => {
@@ -196,5 +105,5 @@ export const bump = async (version: string, options: BumpFlags) => {
 	);
 	const config = await mergeFlags(options);
 	const diff = await getChanged(config);
-	await bumpPackages(diff, config, version);
+	await bumpPackages(diff, version);
 };
